requirements.txt.txt
fastapi
uvicorn
biopython